requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.1
iniconfig==1.1.1
isort==5.10.1
Markdown==3.3.6
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
pathspec==0.9.0
Pillow==5.0.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
zipp==3.7.0